requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
chardet==3.0.4
Django==4.0.1
idna==2.8
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.3.0
requests==2.22.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.25.11
